[code-infra] Fold scope rename into pnpm.mjs and test it through behavior
renameWorkspaceScope sits alongside the other manifest operations, which
lets aliasTarget and renameScope stop being exported: aliasTarget only
lived in pnpm.mjs to keep scope.mjs from importing it circularly.

Drop the aliasTarget and renameScope unit suites. Their cases are covered
where they matter instead: the unscoped alias target and the scope-boundary
rule now ride along in the checkPublishDependencies and renameWorkspaceScope
tests.
